workaround for thinarch
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@46957 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/linux/universal/linux-4.9/arch/mips/mm/tlbex.c
@@ -1645,8 +1645,8 @@ static void build_setup_pgd(void)
 		panic("tlbmiss_handler_setup_pgd space exceeded");
 
 	uasm_resolve_relocs(relocs, labels);
-	pr_debug("Wrote tlbmiss_handler_setup_pgd (%u instructions).\n",
-		 (unsigned int)(p - (u32)&tlbmiss_handler_setup_pgd));
+	pr_debug("Wrote tlbmiss_handler_setup_pgd (%lu instructions).\n",
+		 (size_t)(p - (size_t)&tlbmiss_handler_setup_pgd));
 
 	dump_handler("tlbmiss_handler", (u32*)&tlbmiss_handler_setup_pgd,
 					tlbmiss_handler_setup_pgd_size);
